+++ kafka_message_producer/src/main/java/ru/dankoy/kafkamessageproducer/core/service/messagesender/community/CommunityMessageProducerServiceKafka.java
@@ -1,4 +1,4 @@
-package ru.dankoy.kafkamessageproducer.core.service.messagesender;
+package ru.dankoy.kafkamessageproducer.core.service.messagesender.community;
 
 import java.nio.charset.StandardCharsets;
 import java.util.function.Consumer;
@@ -6,8 +6,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.springframework.kafka.core.KafkaTemplate;
+import ru.dankoy.kafkamessageproducer.core.domain.message.CommunitySubscriptionMessage;
 import ru.dankoy.kafkamessageproducer.core.domain.message.CoubMessage;
-import ru.dankoy.kafkamessageproducer.core.domain.message.TagSubscriptionMessage;
 
 /**
  * @deprecated in favor {@link MessageProducerServiceKafka}
@@ -15,25 +15,27 @@ import ru.dankoy.kafkamessageproducer.core.domain.message.TagSubscriptionMessage
 @Deprecated(since = "2024-02-14")
 @Slf4j
 @RequiredArgsConstructor
-public class TagMessageProducerServiceKafka implements TagMessageProducerService {
+public class CommunityMessageProducerServiceKafka implements CommunityMessageProducerService {
 
   private final KafkaTemplate<String, CoubMessage> kafkaTemplate;
 
   private final String topic;
 
-  private final Consumer<TagSubscriptionMessage> sendAck;
+  private final Consumer<CommunitySubscriptionMessage> sendAck;
 
-  private final Consumer<TagSubscriptionMessage> sendAckToRegistry;
+  private final Consumer<CommunitySubscriptionMessage> sendAckToRegistry;
 
   @Override
-  public void send(TagSubscriptionMessage tagSubscriptionMessage) {
+  public void send(CommunitySubscriptionMessage communitySubscriptionMessage) {
     try {
-      log.info("message: {}", tagSubscriptionMessage);
+      log.info("message: {}", communitySubscriptionMessage);
 
       ProducerRecord<String, CoubMessage> producerRecord =
-          new ProducerRecord<>(topic, tagSubscriptionMessage);
+          new ProducerRecord<>(topic, communitySubscriptionMessage);
 
-      producerRecord.headers().add("subscription_type", "BY_TAG".getBytes(StandardCharsets.UTF_8));
+      producerRecord
+          .headers()
+          .add("subscription_type", "BY_COMMUNITY".getBytes(StandardCharsets.UTF_8));
 
       kafkaTemplate
           .send(producerRecord)
@@ -42,20 +44,19 @@ public class TagMessageProducerServiceKafka implements TagMessageProducerService
                 if (ex == null) {
                   log.info(
                       "message id: {} was sent, offset: {}",
-                      tagSubscriptionMessage.getId(),
+                      communitySubscriptionMessage.getId(),
                       result.getRecordMetadata().offset());
                   // if kafka accepted - send update last permalink in db for
                   // subscription
-                  sendAck.accept(tagSubscriptionMessage);
+                  sendAck.accept(communitySubscriptionMessage);
                   // update registry
-                  sendAckToRegistry.accept(tagSubscriptionMessage);
-                  log.info("acknowledgement sent for {}", tagSubscriptionMessage);
+                  sendAckToRegistry.accept(communitySubscriptionMessage);
                 } else {
-                  log.error("message id:{} was not sent", tagSubscriptionMessage.getId(), ex);
+                  log.error("message id:{} was not sent", communitySubscriptionMessage.getId(), ex);
                 }
               });
     } catch (Exception ex) {
-      log.error("send error, value:{}", tagSubscriptionMessage, ex);
+      log.error("send error, value:{}", communitySubscriptionMessage, ex);
     }
   }
 }
