+++ kafka_message_producer/src/main/java/ru/dankoy/kafkamessageproducer/core/service/messagesender/generic/KafkaTemplateCoubMessage.java
@@ -1,4 +1,4 @@
-package ru.dankoy.kafkamessageproducer.core.service.messagesender;
+package ru.dankoy.kafkamessageproducer.core.service.messagesender.generic;
 
 import org.apache.kafka.clients.producer.ProducerRecord;
 import ru.dankoy.kafkamessageproducer.core.domain.message.CoubMessage;
