+++ kafka_message_producer/src/main/java/ru/dankoy/kafkamessageproducer/core/service/messagesender/tag/TagMessageProducerServiceKafka.java
@@ -1,4 +1,4 @@
-package ru.dankoy.kafkamessageproducer.core.service.messagesender;
+package ru.dankoy.kafkamessageproducer.core.service.messagesender.tag;
 
 import java.nio.charset.StandardCharsets;
 import java.util.function.Consumer;
@@ -6,8 +6,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.springframework.kafka.core.KafkaTemplate;
-import ru.dankoy.kafkamessageproducer.core.domain.message.CommunitySubscriptionMessage;
 import ru.dankoy.kafkamessageproducer.core.domain.message.CoubMessage;
+import ru.dankoy.kafkamessageproducer.core.domain.message.TagSubscriptionMessage;
 
 /**
  * @deprecated in favor {@link MessageProducerServiceKafka}
@@ -15,27 +15,25 @@ import ru.dankoy.kafkamessageproducer.core.domain.message.CoubMessage;
 @Deprecated(since = "2024-02-14")
 @Slf4j
 @RequiredArgsConstructor
-public class CommunityMessageProducerServiceKafka implements CommunityMessageProducerService {
+public class TagMessageProducerServiceKafka implements TagMessageProducerService {
 
   private final KafkaTemplate<String, CoubMessage> kafkaTemplate;
 
   private final String topic;
 
-  private final Consumer<CommunitySubscriptionMessage> sendAck;
+  private final Consumer<TagSubscriptionMessage> sendAck;
 
-  private final Consumer<CommunitySubscriptionMessage> sendAckToRegistry;
+  private final Consumer<TagSubscriptionMessage> sendAckToRegistry;
 
   @Override
-  public void send(CommunitySubscriptionMessage communitySubscriptionMessage) {
+  public void send(TagSubscriptionMessage tagSubscriptionMessage) {
     try {
-      log.info("message: {}", communitySubscriptionMessage);
+      log.info("message: {}", tagSubscriptionMessage);
 
       ProducerRecord<String, CoubMessage> producerRecord =
-          new ProducerRecord<>(topic, communitySubscriptionMessage);
+          new ProducerRecord<>(topic, tagSubscriptionMessage);
 
-      producerRecord
-          .headers()
-          .add("subscription_type", "BY_COMMUNITY".getBytes(StandardCharsets.UTF_8));
+      producerRecord.headers().add("subscription_type", "BY_TAG".getBytes(StandardCharsets.UTF_8));
 
       kafkaTemplate
           .send(producerRecord)
@@ -44,19 +42,20 @@ public class CommunityMessageProducerServiceKafka implements CommunityMessagePro
                 if (ex == null) {
                   log.info(
                       "message id: {} was sent, offset: {}",
-                      communitySubscriptionMessage.getId(),
+                      tagSubscriptionMessage.getId(),
                       result.getRecordMetadata().offset());
                   // if kafka accepted - send update last permalink in db for
                   // subscription
-                  sendAck.accept(communitySubscriptionMessage);
+                  sendAck.accept(tagSubscriptionMessage);
                   // update registry
-                  sendAckToRegistry.accept(communitySubscriptionMessage);
+                  sendAckToRegistry.accept(tagSubscriptionMessage);
+                  log.info("acknowledgement sent for {}", tagSubscriptionMessage);
                 } else {
-                  log.error("message id:{} was not sent", communitySubscriptionMessage.getId(), ex);
+                  log.error("message id:{} was not sent", tagSubscriptionMessage.getId(), ex);
                 }
               });
     } catch (Exception ex) {
-      log.error("send error, value:{}", communitySubscriptionMessage, ex);
+      log.error("send error, value:{}", tagSubscriptionMessage, ex);
     }
   }
 }
